feat: farm bot contract from hackathon (PR 1/3) (#1)
Plus some tooling.

The main value add here is extracting this from a rather messy hackathon submission repo.

I also added the "deployments" directory, which I feel less strongly about but it seems like a nice record of what contracts we have out there.

also included a fix for first-time compilation (previously there was a circular dependency in types, where deploy.ts required typechain types, but typechain could not be run without compilation errors in deploy.ts). The fix was to use a more general type for the deploy function and swallow the errors with ts-ignore.