fix: avoid safeApprove (#55)
it's deprecated, and fails when setting to something nonzero after an initial approval

I changed it to safeIncreaseAllowance in the place we are guaranteed to spend the increased allowance straight after. but in the place where we then deposit into a liquidity pool that could lead to an unspent allowance, so I used a regular approve there. The other option would be to set the approval to 0 right after depositing into a liquidity pool, but I think that's over-doing it. Competing products (Beefy, looking at you) approve the router for infinite amounts to save gas.